//Added new gitignore file //Added test for chapter1 page
