+++ .gitignore
@@ -1,0 +1,3 @@
+/target/
+/selenium-learning-tb-book-example-2020.iml
+/.idea/
